--- a/src/Avalonia.Base/AvaloniaObject.cs
+++ b/src/Avalonia.Base/AvaloniaObject.cs
@@ -5,6 +5,7 @@ using System.Runtime.CompilerServices;
 using Avalonia.Data;
 using Avalonia.Diagnostics;
 using Avalonia.Logging;
+using Avalonia.Media;
 using Avalonia.PropertyStore;
 using Avalonia.Threading;
 
@@ -20,7 +21,7 @@ namespace Avalonia
     {
         private readonly ValueStore _values;
         private AvaloniaObject? _inheritanceParent;
-        private List<AvaloniaObject>? _mediaParents;
+        private MediaParentsBag<AvaloniaObject>? _mediaParents;
         private PropertyChangedEventHandler? _inpcChanged;
         private EventHandler<AvaloniaPropertyChangedEventArgs>? _propertyChanged;
         private List<AvaloniaObject>? _inheritanceChildren;
@@ -80,8 +81,8 @@ namespace Avalonia
             }
         }
 
-        internal List<AvaloniaObject> GetOrCreateMediaParents() => _mediaParents ??= new();
-        internal List<AvaloniaObject>? GetMediaParents() => _mediaParents;
+        internal MediaParentsBag<AvaloniaObject> GetOrCreateMediaParents() => _mediaParents ??= new();
+        internal MediaParentsBag<AvaloniaObject>? GetMediaParents() => _mediaParents;
 
         /// <summary>
         /// Gets or sets the value of a <see cref="AvaloniaProperty"/>.