--- a/src/Avalonia.Base/Media/Brush.cs
+++ b/src/Avalonia.Base/Media/Brush.cs
@@ -111,6 +111,8 @@ namespace Avalonia.Media
             {
                 property.Changed.Subscribe(invalidateObserver);
             }
+
+            MediaInvalidation.AffectsMediaRender(properties);
         }
 
         /// <summary>