--- a/src/Avalonia.Base/Media/DrawingCollection.cs
+++ b/src/Avalonia.Base/Media/DrawingCollection.cs
@@ -3,16 +3,14 @@ using Avalonia.Collections;
 
 namespace Avalonia.Media
 {
-    public sealed class DrawingCollection : AvaloniaList<Drawing>
+    public sealed class DrawingCollection : MediaCollection<Drawing>
     {
         public DrawingCollection()
         {
-            ResetBehavior = ResetBehavior.Remove;
         }
 
         public DrawingCollection(IEnumerable<Drawing> items) : base(items)
         {
-            ResetBehavior = ResetBehavior.Remove;
         }
     }
 }