--- a/src/Avalonia.Base/Media/Geometry.cs
+++ b/src/Avalonia.Base/Media/Geometry.cs
@@ -127,6 +127,8 @@ namespace Avalonia.Media
             {
                 property.Changed.Subscribe(invalidateObserver);
             }
+
+            MediaInvalidation.AffectsMediaRender(properties);
         }
 
         /// <summary>