--- a/src/Avalonia.Base/Media/GeometryCollection.cs
+++ b/src/Avalonia.Base/Media/GeometryCollection.cs
@@ -1,45 +1,35 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using Avalonia.Collections;
 
 #nullable enable
 
 namespace Avalonia.Media
 {
-    public sealed class GeometryCollection : AvaloniaList<Geometry> 
+    public sealed class GeometryCollection : MediaCollection<Geometry>
     {
         public GeometryCollection()
         {
-            ResetBehavior = ResetBehavior.Remove;
-
-            this.ForEachItem(
-               x =>
-               {
-                   Parent?.Invalidate();
-               },
-               x =>
-               {
-                   Parent?.Invalidate();
-               },
-               () => throw new NotSupportedException());
+            Setup();
         }
 
         public GeometryCollection(IEnumerable<Geometry> items) : base(items)
         {
-            ResetBehavior = ResetBehavior.Remove;
-
-            this.ForEachItem(
-               x =>
-               {
-                   Parent?.Invalidate();
-               },
-               x =>
-               {
-                   Parent?.Invalidate();
-               },
-               () => throw new NotSupportedException());
+            Setup();
         }
 
-        public GeometryGroup? Parent { get; set; }
+        private void Setup()
+        {
+            this.ForEachItem(Invalidate, Invalidate, () => throw new NotSupportedException());
+        }
+
+        private void Invalidate(Geometry child)
+        {
+            foreach (var parent in Parents.OfType<GeometryGroup>())
+            {
+                parent.Invalidate();
+            }
+        }
     }
 }