--- a/src/Avalonia.Base/Media/GlyphRunDrawing.cs
+++ b/src/Avalonia.Base/Media/GlyphRunDrawing.cs
@@ -20,6 +20,11 @@
             set => SetValue(GlyphRunProperty, value);
         }
 
+        static GlyphRunDrawing()
+        {
+            MediaInvalidation.AffectsMediaRender(ForegroundProperty, GlyphRunProperty);
+        }
+
         public override void Draw(DrawingContext context)
         {
             if (GlyphRun == null)