--- a/src/Avalonia.Base/Media/MediaCollection.cs
+++ b/src/Avalonia.Base/Media/MediaCollection.cs
@@ -2,15 +2,26 @@
 using System.Collections;
 using System.Collections.Generic;
 using Avalonia.Collections;
+using Avalonia.Metadata;
 
 #nullable enable
 
 namespace Avalonia.Media
 {
+    /// <summary>
+    /// Base type for collections of objects that participate in <see cref="MediaInvalidation"/>.
+    /// </summary>
     public abstract class MediaCollection<T> : AvaloniaList<T>, IMediaCollection where T : AvaloniaObject
     {
         private readonly MediaParentsBag<AvaloniaObject> _parentHandles = new();
 
+        /// <summary>
+        /// Gets an enumeration of living media parents of this collection.
+        /// </summary>
+        /// <remarks>
+        /// To invoke changes to this property, assign the <see cref="MediaCollection{T}"/> to an <see cref="AvaloniaProperty"/> 
+        /// which has been passed to <see cref="MediaInvalidation.AffectsMediaRender"/>.
+        /// </remarks>
         public IEnumerable<AvaloniaObject> Parents => _parentHandles;
 
         protected MediaCollection()