… src/Avalonia.Base/Media/MediaParentsBag.cs
… src/Avalonia.Base/Media/PathGeometry.cs
… src/Avalonia.Base/Media/PathGeometryCollections.cs
… src/Avalonia.Base/Media/Pen.cs
… src/Avalonia.Base/Media/PolylineGeometry.cs
… src/Avalonia.Base/Visual.cs
… src/Avalonia.Controls/Shapes/Shape.cs
… tests/Avalonia.Controls.UnitTests/MediaInvalidationTests.cs