@kekekeks 2022-09-06
The general plan was to move media invalidation to the render thread by introducing render-thread representations for media objects (the same way it's done in UWP). That would also remove the need to call `Render` and re-build the list of drawing commands.