@kekekeks 2022-09-06
`MediaParentsBag` and `MediaCollection` look like a good design in general, I think those can be adapted for use on the render thread.

However I'm not sure that we want to merge this PR as is if we are going to implement CompositionBrush, CompositionShape and friends, since in that case we'll have to remove the changes of this PR later.

I'd definitely would like to merge this if we won't get composition objects to work in time before the release.


The general roadmap is:
- remove DeferrredRenderer/ImmediateRenderer, so that Compositor is always used
- change Visual to synchronize properties instead of calling InvalidateVisual all the time.
- implement media objects that live on the render thread

I hope to get that done before November this year.


BTW, there is also a similar problem with UI-thread animations: right now they don't track visual tree attachment at all, so if you have some brush with animation in resources, it would tick even if brush isn't used anywhere.