@Whiletru3 2023-02-07
Any chance to backport this PR in the stable 0.10.x branch ?

We are facing this issue with drawing images used as icons in out desktop application. Those are dependant of a Material.Avalonia color theme MaterialDesignBody.

When we change the theme from light to dark (or vice versa), the images are not updated.
Example :
```
<DrawingGroup x:Key="AddBookmarkDrawing">
        <DrawingGroup.Children>
            <GeometryDrawing Geometry="F1 M 16,9.98975L 5.98291,9.98975L 5.98291,26.0171L 22.0103,26.0171L 22.0103,16">
                <GeometryDrawing.Pen>
                    <Pen Thickness="1.04178" LineJoin="Round" Brush="{DynamicResource MaterialDesignBody}" />
                </GeometryDrawing.Pen>
            </GeometryDrawing>
            <GeometryDrawing Geometry="F1 M 18.0034,9.98975L 26.0171,9.98975">
                <GeometryDrawing.Pen>
                    <Pen Thickness="1.04178" LineJoin="Round" Brush="{DynamicResource MaterialDesignBody}" />
                </GeometryDrawing.Pen>
            </GeometryDrawing>
            <GeometryDrawing Geometry="F1 M 22.0103,5.98291L 22.0103,13.9966">
                <GeometryDrawing.Pen>
                    <Pen Thickness="1.04178" LineJoin="Round" Brush="{DynamicResource MaterialDesignBody}" />
                </GeometryDrawing.Pen>
            </GeometryDrawing>
            <GeometryDrawing Brush="{DynamicResource MaterialDesignBody}"
                             Geometry="F1 M 15.5993,14.5575L 12.2336,14.5575C 11.9932,14.5575 11.6726,14.8781 11.6726,15.1986L 11.6726,21.3692L 13.9966,19.3657L 16.2404,21.3692L 16.2404,15.1986C 16.1603,14.8781 15.8397,14.5575 15.5993,14.5575 Z " />
        </DrawingGroup.Children>
    </DrawingGroup>

    <DrawingImage x:Key="AddBookmarkImage" Drawing="{DynamicResource AddBookmarkDrawing}" />
```