@kekekeks 2023-02-07
Since compositor-side brushes aren't likely to land in 11.0, we are likely to be using the approach from this PR for the time being.