@kekekeks 2023-05-15
Brushes/pens are now automatically updating render-thread resources - https://github.com/AvaloniaUI/Avalonia/pull/11340

Geometries will use the same infrastructure later.