@robloo 2023-05-17
@tomenscape I saw #11378 was opened for this. Doesnt look like it is planned for 11.0 though...