@kekekeks 2024-06-19
We've discontinued IAffectsRender usage. Every resource is responsible of updating its server counterpart by itself.